A macro implemented “template” for a double-ended queue container,
known as a deque.
The Deque is implemented as a circular, dynamic array.
The following is the way to create a Deque of integer type which can be
used in your code:

Deque_DEFINE(int)

This will basically result in a text replacement where
Deque_DEFINE(int) will be replace the code for Deque of type integer.

The struct for a container of int variables would be named Deque_int.